requirements.txt
typing-extensions==4.1.1
python-multipart==0.0.5
numpy==1.19.4
opencv-python==4.5.4.60
pandas==1.1.5
Pillow==8.3.2
fastapi==0.75.2
click==7.1.2
asgiref==3.4.1
uvicorn==0.16.0
